requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
huggingface-hub>=0.20.0
requests==2.31.0